requirements.txt
fastapi
uvicorn
pydantic
python-multipart
passlib[bcrypt]
python-jose
pymongo
python-dotenv
json5
